requirements.txt
# RCE-LLM Production Requirements
# Author: Ismail Sialyen
# Paper: DOI 10.5281/zenodo.17360372

# Core NLP
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Optimization (optional but recommended)
pulp>=2.7.0          # For ILP optimization
scipy>=1.11.0        # For numerical optimization

# Date/time parsing
python-dateutil>=2.8.2

# Data structures
networkx>=3.1        # For graph analysis

# Scientific computing
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0        # Code formatting
flake8>=6.1.0        # Linting
mypy>=1.5.0          # Type checking

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# API (for production deployment)
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0

# Retrieval (for RAG baseline)
sentence-transformers>=2.2.2  # For embeddings
faiss-cpu>=1.7.4              # For vector search

# Benchmarking
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
